# back-end/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Dhumi Demo Production API
  version: 0.1.0
  description: |
    Accepted mock-first implementation contract for the Dhumi Marketplace Dataset
    and Scraper Library demo. Customers call only Dhumi. Bright Data credentials,
    dataset IDs, snapshot IDs and raw errors are never public fields.
    Customer operations use browser sessions. The 0070 refactor removes the
    three /v1/keys operations and secret-response recovery.

    This contract has no activation, customer plan, subscription, billing,
    credit or quota endpoints. Organization, verification and member workflows follow 0071.
    Live Bright Data traffic remains controlled by external evidence gates.
  contact:
    name: Dhumi Engineering
  license:
    name: Proprietary - internal implementation contract
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.dhumi.example
    description: Reserved documentation placeholder; deployment base URL is pending.
...
paths:
...
  /v1/organizations:
    get:
      tags:
        - Organizations
      operationId: listOrganizations
      x-dhumi-access: browser_session
      summary: list Organizations
      description: Organization workflows are implemented for 0071; database cutover and qualification remain pending.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Organizations
      operationId: createOrganization
      x-dhumi-access: browser_session
      summary: create Organization
      description: Organization workflows are implemented for 0071; database cutover and qualification remain pending.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreateInput'
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/invites/accept:
    post:
      tags:
        - Organizations
      operationId: acceptInvite
      x-dhumi-access: browser_session
      summary: accept Invite
      description: >-
        Secrets appear only on the first minting response. An idempotent replay returns metadata; explicit resend rotates
        the token. List responses never reveal token hashes or secrets.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteAcceptInput'
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/password-reset:
    post:
      tags:
        - Organizations
      operationId: requestPasswordReset
      x-dhumi-access: public
      summary: request Password Reset
      description: Organization workflows are implemented for 0071; database cutover and qualification remain pending.
      security: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetInput'
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/verifications/{verification_id}/confirm:
    post:
      tags:
        - Organizations
      operationId: confirmVerification
      x-dhumi-access: purpose_bound_verification
      summary: confirm Verification
      description: >-
        Create/join requires the same signed-in user and CSRF; reset requires code and new_password, no email-link token.
        Fresh proof expires in ten minutes; five wrong attempts end it. A resend rotates code/link, with a 60-second
        cooldown and five issues per hour across purposes. GET never consumes a proof.
      security:
        - {}
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: verification_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: X-CSRF-Token
          in: header
          required: false
          schema:
            type: string
          description: Required for signed-in create/join proofs; reset is public OTP-only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationConfirmInput'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationConfirmed'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/verifications/{verification_id}/resend:
    post:
      tags:
        - Organizations
      operationId: resendVerification
      x-dhumi-access: purpose_bound_verification
      summary: resend Verification
      description: >-
        Create/join requires the same signed-in user and CSRF; reset requires code and new_password, no email-link token.
        Fresh proof expires in ten minutes; five wrong attempts end it. A resend rotates code/link, with a 60-second
        cooldown and five issues per hour across purposes. GET never consumes a proof.
      security:
        - {}
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: verification_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: X-CSRF-Token
          in: header
          required: false
          schema:
            type: string
          description: Required for signed-in create/join proofs; reset is public OTP-only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyWorkflowInput'
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/organization/members:
    get:
      tags:
        - Organizations
      operationId: listMembers
      x-dhumi-access: organization_member
      summary: list Organization Members
      description: Organization workflows are implemented for 0071; database cutover and qualification remain pending.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/OrganizationSelector'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationMemberPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/organization/members/{user_id}:
    patch:
      tags:
        - Organizations
      operationId: updateMember
      x-dhumi-access: organization_admin
      summary: change Organization Member
      description: Organization workflows are implemented for 0071; database cutover and qualification remain pending.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/OrganizationSelector'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationMemberChange'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationActionCompleted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Organizations
      operationId: removeMember
      x-dhumi-access: organization_admin
      summary: remove Organization Member
      description: Organization workflows are implemented for 0071; database cutover and qualification remain pending.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/OrganizationSelector'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Completed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/organization/invites:
    get:
      tags:
        - Organizations
      operationId: listInvites
      x-dhumi-access: organization_admin
      summary: list Organization Invites
      description: >-
        Secrets appear only on the first minting response. An idempotent replay returns metadata; explicit resend rotates
        the token. List responses never reveal token hashes or secrets.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/OrganizationSelector'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationInvitePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Organizations
      operationId: createInvite
      x-dhumi-access: organization_admin
      summary: create Organization Invite
      description: >-
        Secrets appear only on the first minting response. An idempotent replay returns metadata; explicit resend rotates
        the token. List responses never reveal token hashes or secrets.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/OrganizationSelector'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationInviteInput'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationInviteMinted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/organization/invites/{invite_id}:
    delete:
      tags:
        - Organizations
      operationId: revokeInvite
      x-dhumi-access: organization_admin
      summary: revoke Organization Invite
      description: >-
        Secrets appear only on the first minting response. An idempotent replay returns metadata; explicit resend rotates
        the token. List responses never reveal token hashes or secrets.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: invite_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/OrganizationSelector'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Completed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/organization/invites/{invite_id}/resend:
    post:
      tags:
        - Organizations
      operationId: resendInvite
      x-dhumi-access: organization_admin
      summary: resend Organization Invite
      description: >-
        Secrets appear only on the first minting response. An idempotent replay returns metadata; explicit resend rotates
        the token and ensures at least the configured backend resend lifetime remains (seven days by default), without
        shortening a later expiry. Revoked, expired or exhausted invitations remain unavailable. List responses never
        reveal token hashes or secrets. Delivery is attempted after commit; a delivery failure still returns 202.
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: invite_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/OrganizationSelector'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyWorkflowInput'
      responses:
        '202':
          description: Invitation rotated; delivery attempted after commit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationInviteMinted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BrowserBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived Dhumi browser access token.
...
  parameters:
...
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller correlation ID; Dhumi returns an accepted/generated ID.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Opaque caller key scoped by authenticated actor/Tenant and operation.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Required for state-changing browser session operations.
      schema:
        type: string
        minLength: 16
        maxLength: 512
...
  schemas:
...
    VerificationAccepted:
      type: object
      additionalProperties: false
      required:
        - accepted
        - verification_id
      properties:
        accepted:
          type: boolean
          const: true
        verification_id:
          type: string
          format: uuid
        message:
          type: string
...
    OrganizationPage:
      type: object
      additionalProperties: false
      required:
        - organizations
      properties:
        organizations:
          type: array
          items:
            $ref: '#/components/schemas/OrganizationSummary'
    OrganizationCreateInput:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          pattern: \S
    InviteAcceptInput:
      type: object
      additionalProperties: false
      required: []
      properties:
        join_code:
          type: string
          pattern: ^[A-Za-z0-9_-]{43}$
        invite_token:
          type: string
          pattern: ^[A-Za-z0-9_-]{43}$
      oneOf:
        - required:
            - join_code
        - required:
            - invite_token
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PayloadTooLarge:
      description: Request body exceeds the accepted one-mebibyte limit
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnsupportedMediaType:
      description: Request body does not use a supported media type
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid Dhumi authentication
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but not authorized or Tenant unavailable
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource does not exist or is not visible to this Tenant
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Idempotency or current-state conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: Semantically invalid request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TooManyRequests:
      description: Temporary Dhumi platform/product capacity limit; not a customer quota
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InternalServerError:
      description: Unexpected Dhumi server error; internal detail is never exposed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...